requirements.txt
asgiref==3.7.2
asttokens==2.4.1
black==23.12.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
decorator==5.1.1
defusedxml==0.7.1
Django==5.0
django-allauth==0.59.0
exceptiongroup==1.2.0
executing==2.0.1
facebook-sdk==3.1.0
idna==3.6
ipython==8.19.0
jedi==0.19.1
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.2
parso==0.8.3
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
stack-data==0.6.3
tomli==2.0.1
traitlets==5.14.0
tweepy==4.14.0
typing_extensions==4.9.0
urllib3==2.1.0
wcwidth==0.2.12
